+++ populationRequest/store/indexPopulationRequest.js
@@ -1,3 +1,4 @@
+import actions from "./actionsPopulationRequest";
 import getters from "./gettersPopulationRequest";
 import mutations from "./mutationsPopulationRequest";
 import state from "./statePopulationRequest";
@@ -6,5 +7,6 @@ export default {
     namespaced: true,
     state: {...state},
     mutations,
-    getters
+    getters,
+    actions
 };